Crate hudhook

hudhook

This library implements a mechanism for hooking into the render loop of DirectX11 applications, perform memory manipulation and draw things on screen via imgui. It has been largely inspired by CheatEngine.

It's been extracted out of the darksoulsiii-practice-tool for generalized usage as a stand-alone framework. It is also a complete, fully-fledged example of usage; it is a good idea to refer to that for any doubts about the API which aren't clarified by this documentation.

Refer to this post for in-depth information about the architecture of the library.

Fair warning

hudhook provides essential, crash-safe features for memory manipulation and UI rendering. It does, alas, contain a hefty amount of FFI and unsafe code which still has to be thoroughly tested, validated and audited for soundness. It should be OK for small projects such as videogame mods, but it may crash your application at this stage.

Examples

Hooking the render loop and drawing things with imgui:

Compile your crate with both a cdylib and an executable target. The executable will be very minimal and used to inject the DLL into the target process.

// lib.rs
use hudhook::*;

pub struct MyRenderLoop;
impl RenderLoop for MyRenderLoop {
  fn render(&self, ctx: hudhook::RenderContext) {
   imgui::Window::new(im_str!("My first render loop"))
    .position([0., 0.], imgui::Condition::FirstUseEver)
    .size([320., 200.], imgui::Condition::FirstUseEver)
    .build(ctx.frame, || {
      ctx.frame.text(imgui::im_str!("Hello, hello!"));
    });
  }

  fn is_visible(&self) -> bool { true }
  fn is_capturing(&self) -> bool { true }
}

hudhook!(Box::new(MyRenderLoop::new()))

// main.rs
use hudhook::inject;

fn main() {
  let mut cur_exe = std::env::current_exe().unwrap();
  cur_exe.push("..");
  cur_exe.push("libmyhook.dll");

  let cur_dll = cur_exe.canonicalize().unwrap();

  inject("MyTargetApplication.exe", cur_dll.as_path().to_str().unwrap()).unwrap();
}

Memory manipulation

In an initialization step:

let x = PointerChain::<f32>::new(&[base_address, 0x40, 0x28, 0x80]);
let y = PointerChain::<f32>::new(&[base_address, 0x40, 0x28, 0x88]);
let z = PointerChain::<f32>::new(&[base_address, 0x40, 0x28, 0x84]);

In the render loop:

x.read().map(|val| x.write(val + 1.));
y.read().map(|val| y.write(val + 1.));
z.read().map(|val| z.write(val + 1.));

Re-exports

pub use imgui;

pub use winapi;

Modules

memory
Memory manipulation API.

Macros

hudhook
Entry point for the library.

Structs

Error
Trivial to-string implementation of an error.

RenderContext
Information context made available to the RenderLoop

Traits

RenderLoop
Interface for implementing the render loop.

Functions

apply_hook
Inner entry point for the library.

get_dll_path
Returns the path of the implementor's DLL.

inject
Inject dll inside the process_name process.