Function hudhook::inject

pub fn inject(process_name: &str, dll: &str) -> Result<(), Error>

Inject dll inside the process_name process.

Create a bin target for your mod and use this function to build an executable which launches your mod's dll.

Example:

use hudhook::prelude::inject;

pub fn main() {
  inject("DarkSoulsIII.exe", "darksoulsiii-practice-tool.dll");
}