Trait hudhook::RenderLoop

pub trait RenderLoop {
    fn render<'a>(&mut self, ctx: RenderContext);

    fn is_visible(&self) -> bool;

    fn is_capturing(&self) -> bool;
}

Interface for implementing the render loop.

Required methods

fn render<'a>(&mut self, ctx: RenderContext)

Invoked once per frame. Memory management and UI visualization (via the current frame's imgui::Ui instance) should be made inside of it.

fn is_visible(&self) -> bool

Return true when you want your UI to be rendered to screen.

The render method will still be called, but the draw data will not be displayed

fn is_capturing(&self) -> bool

Return true when you want the underlying application to stop receiving WndProc events. Presently not functioning.

Loading content...